% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{adstock}
\alias{adstock}
\title{adstock smoothing function}
\usage{
adstock(x, weights, L, args, na.rm = T)
}
\arguments{
\item{x}{numeric vector. It indicates the amount of spendings, all of which elements must be non-negative.}

\item{weights}{function. typically, you can use \link{get_weight} function.}

\item{L}{integer. maximum lag of carryover effect.}

\item{args}{list. extra arguments for \code{weights} function.}

\item{na.rm}{logical. default is \code{TRUE}.}
}
\description{
adstock smoothing function
}
