% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{geo_weight}
\alias{geo_weight}
\title{geometric weight function}
\usage{
geo_weight(l, alpha, theta)
}
\arguments{
\item{l}{integer. lag index.}

\item{alpha}{numeric. It means the decay rate parameter, which must be larger than zero (0) and less than one (1).}

\item{theta}{numeric. It means the delay parameter, which must be larger than or equatl zero (0) and less than or equalt maximum lag minus one (1).}
}
\description{
used for \link{adstock} function
}
