% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{hill}
\alias{hill}
\title{hill function}
\usage{
hill(x, k, s)
}
\arguments{
\item{x}{numeric vector. It indicates the amount of spendings, all of which elements must be non-negative.}

\item{k}{numeric. It means the half-satuation point parameter, which must be non-negative.}

\item{s}{numeric. It means the \link{s}lope or \link{s}hape parameter, which must be non-negative.}
}
\description{
hill function
}
