% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{predict}
\alias{predict}
\title{Return predict values from MMM stan fit object}
\usage{
predict(fit, newdata = NULL)
}
\arguments{
\item{fit}{stanfit object.}

\item{newdata}{same as stan input list. if not specified, return predicted values for fitting data.}
}
\value{
matrix, which has (iter x chain) rows and (Times) columns.
}
\description{
Return predict values from MMM stan fit object
}
