% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ROAS}
\alias{ROAS}
\title{Calculate (Marginal) Return On the Advertising Spendings}
\usage{
ROAS(fit, data = NULL, L, controls = "price", reference = 0)
}
\arguments{
\item{fit}{model parameter object}

\item{data}{data for extraporation}

\item{reference}{numeric reference levels for spendings. default is zero.}
}
\description{
Calculate (Marginal) Return On the Advertising Spendings
}
