% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{to_stan_data}
\alias{to_stan_data}
\title{convert wider data frame to stan input object}
\usage{
to_stan_data(data, L, controls = "price")
}
\arguments{
\item{data}{raw data}

\item{L}{max lag.}

\item{controls}{default \code{"price"}}
}
\description{
convert wider data frame to stan input object
}
