% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ROAS}
\alias{ROAS}
\title{Calculate (Marginal) Return On the Advertising Spendings}
\usage{
ROAS(
  fit,
  data = NULL,
  L,
  controls = "price",
  reference = 0,
  inverse_trans = NULL
)
}
\arguments{
\item{fit}{model parameter object}

\item{data}{data for extraporation}

\item{reference}{numeric reference levels for spendings. default is zero.}

\item{inverse_trans}{function. inverse transformation applying for the response variable}
}
\description{
Calculate (Marginal) Return On the Advertising Spendings
}
